+++ INTERX/common/api.go
@@ -159,7 +159,7 @@ func GetAccountNumberSequence(gwCosmosmux *runtime.ServeMux, r *http.Request, be
 
 // BroadcastTransaction is a function to post transaction, returns txHash
 func BroadcastTransaction(rpcAddr string, txBytes []byte) (string, error) {
-	endpoint := fmt.Sprintf("%s/broadcast_tx_commit?tx=0x%X", rpcAddr, txBytes)
+	endpoint := fmt.Sprintf("%s/broadcast_tx_async?tx=0x%X", rpcAddr, txBytes)
 	GetLogger().Info("[rpc-call] Entering rpc call: ", endpoint)
 
 	resp, err := http.Get(endpoint)
@@ -191,11 +191,6 @@ func BroadcastTransaction(rpcAddr string, txBytes []byte) (string, error) {
 	if resp.StatusCode != http.StatusOK {
 		GetLogger().Error("[rpc-call] Unable to broadcast transaction: ", result.Error.Message)
 		return "", errors.New(result.Error.Message)
-	}
-
-	if result.Result.Height == "0" {
-		GetLogger().Error("[rpc-call] Failed to broadcast transaction")
-		return "", fmt.Errorf("failed to send tokens")
 	}
 
 	return result.Result.Hash, nil
@@ -335,6 +330,31 @@ func GetTokenAliases(gwCosmosmux *runtime.ServeMux, r *http.Request) []types.Tok
 	database.AddTokenAliases(result.Data)
 
 	return result.Data
+}
+
+// GetTokenSupply is a function to get token supply
+func GetTokenSupply(gwCosmosmux *runtime.ServeMux, r *http.Request) []types.TokenSupply {
+	r.URL.Path = config.QueryTotalSupply
+	r.URL.RawQuery = ""
+	r.Method = "GET"
+
+	GetLogger().Info("[grpc-call] Entering grpc call: ", r.URL.Path)
+
+	recorder := httptest.NewRecorder()
+	gwCosmosmux.ServeHTTP(recorder, r)
+	resp := recorder.Result()
+
+	type TokenAliasesResponse struct {
+		Supply []types.TokenSupply `json:"supply"`
+	}
+
+	result := TokenAliasesResponse{}
+	err := json.NewDecoder(resp.Body).Decode(&result)
+	if err != nil {
+		GetLogger().Error("[grpc-call] Unable to decode response: ", err)
+	}
+
+	return result.Supply
 }
 
 func GetKiraStatus(rpcAddr string) *types.KiraStatus {