+++ INTERX/common/hash.go
@@ -2,16 +2,15 @@ package common
 
 import (
 	"crypto/md5"
+	"crypto/sha256"
 	"encoding/json"
 	"fmt"
-
-	"golang.org/x/crypto/blake2b"
 )
 
-// GetBlake2bHashFromBytes is a function to get hash
-func GetBlake2bHashFromBytes(data []byte) string {
-	hash := blake2b.Sum256(data)
-	return fmt.Sprintf("%X", hash)
+// GetSha256SumFromBytes is a function to get hash
+func GetSha256SumFromBytes(data []byte) string {
+	hash := sha256.Sum256(data)
+	return fmt.Sprintf("%x", hash)
 }
 
 // GetBlake2bHash is a function to get hash
@@ -22,7 +21,7 @@ func GetBlake2bHash(request interface{}) string {
 		GetLogger().Error("[blake2b-hash] Unable to marshal request: ", err)
 	}
 
-	return GetBlake2bHashFromBytes(requestJSON)
+	return GetSha256SumFromBytes(requestJSON)
 }
 
 // GetMD5Hash is a function to get hash