+++ simapp/app.go
@@ -198,7 +198,7 @@ func NewSimApp(
 
 	app.CustomGovKeeper = customgovkeeper.NewKeeper(keys[customgovtypes.ModuleName], appCodec)
 	customStakingKeeper := keeper.NewKeeper(keys[customstakingtypes.ModuleName], legacyAmino, app.CustomGovKeeper)
-	app.CustomSlashingKeeper = customslashingkeeper.NewKeeper(appCodec, keys[customslashingtypes.ModuleName], &customStakingKeeper, app.GetSubspace(customslashingtypes.ModuleName))
+	app.CustomSlashingKeeper = customslashingkeeper.NewKeeper(appCodec, keys[customslashingtypes.ModuleName], &customStakingKeeper, app.CustomGovKeeper, app.GetSubspace(customslashingtypes.ModuleName))
 	app.TokensKeeper = tokenskeeper.NewKeeper(keys[tokenstypes.ModuleName], appCodec)
 	app.CustomStakingKeeper = *customStakingKeeper.SetHooks(
 		customstakingtypes.NewMultiStakingHooks(app.CustomSlashingKeeper.Hooks()),