+++ x/slashing/keeper/hooks.go
@@ -17,7 +17,6 @@ func (k Keeper) AfterValidatorJoined(ctx sdk.Context, address sdk.ConsAddress, v
 		signingInfo := types.NewValidatorSigningInfo(
 			address,
 			ctx.BlockHeight(),
-			0,
 			time.Unix(0, 0),
 			false,
 			0,