+++ x/slashing/types/genesis.pb.go
@@ -30,9 +30,6 @@ type GenesisState struct {
 	// signing_infos represents a map between validator addresses and their
 	// signing infos.
 	SigningInfos []SigningInfo `protobuf:"bytes,2,rep,name=signing_infos,json=signingInfos,proto3" json:"signing_infos" yaml:"signing_infos"`
-	// signing_infos represents a map between validator addresses and their
-	// missed blocks.
-	MissedBlocks []ValidatorMissedBlocks `protobuf:"bytes,3,rep,name=missed_blocks,json=missedBlocks,proto3" json:"missed_blocks" yaml:"missed_blocks"`
 }
 
 func (m *GenesisState) Reset()         { *m = GenesisState{} }
@@ -78,13 +75,6 @@ func (m *GenesisState) GetParams() Params {
 func (m *GenesisState) GetSigningInfos() []SigningInfo {
 	if m != nil {
 		return m.SigningInfos
-	}
-	return nil
-}
-
-func (m *GenesisState) GetMissedBlocks() []ValidatorMissedBlocks {
-	if m != nil {
-		return m.MissedBlocks
 	}
 	return nil
 }
@@ -144,154 +134,35 @@ func (m *SigningInfo) GetValidatorSigningInfo() ValidatorSigningInfo {
 	return ValidatorSigningInfo{}
 }
 
-// ValidatorMissedBlocks contains array of missed blocks of corresponding
-// address.
-type ValidatorMissedBlocks struct {
-	// address is the validator address.
-	Address string `protobuf:"bytes,1,opt,name=address,proto3" json:"address,omitempty"`
-	// missed_blocks is an array of missed blocks by the validator.
-	MissedBlocks []MissedBlock `protobuf:"bytes,2,rep,name=missed_blocks,json=missedBlocks,proto3" json:"missed_blocks" yaml:"missed_blocks"`
-}
-
-func (m *ValidatorMissedBlocks) Reset()         { *m = ValidatorMissedBlocks{} }
-func (m *ValidatorMissedBlocks) String() string { return proto.CompactTextString(m) }
-func (*ValidatorMissedBlocks) ProtoMessage()    {}
-func (*ValidatorMissedBlocks) Descriptor() ([]byte, []int) {
-	return fileDescriptor_14205810582f3203, []int{2}
-}
-func (m *ValidatorMissedBlocks) XXX_Unmarshal(b []byte) error {
-	return m.Unmarshal(b)
-}
-func (m *ValidatorMissedBlocks) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	if deterministic {
-		return xxx_messageInfo_ValidatorMissedBlocks.Marshal(b, m, deterministic)
-	} else {
-		b = b[:cap(b)]
-		n, err := m.MarshalToSizedBuffer(b)
-		if err != nil {
-			return nil, err
-		}
-		return b[:n], nil
-	}
-}
-func (m *ValidatorMissedBlocks) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_ValidatorMissedBlocks.Merge(m, src)
-}
-func (m *ValidatorMissedBlocks) XXX_Size() int {
-	return m.Size()
-}
-func (m *ValidatorMissedBlocks) XXX_DiscardUnknown() {
-	xxx_messageInfo_ValidatorMissedBlocks.DiscardUnknown(m)
-}
-
-var xxx_messageInfo_ValidatorMissedBlocks proto.InternalMessageInfo
-
-func (m *ValidatorMissedBlocks) GetAddress() string {
-	if m != nil {
-		return m.Address
-	}
-	return ""
-}
-
-func (m *ValidatorMissedBlocks) GetMissedBlocks() []MissedBlock {
-	if m != nil {
-		return m.MissedBlocks
-	}
-	return nil
-}
-
-// MissedBlock contains height and missed status as boolean.
-type MissedBlock struct {
-	// index is the height at which the block was missed.
-	Index int64 `protobuf:"varint,1,opt,name=index,proto3" json:"index,omitempty"`
-	// missed is the missed status.
-	Missed bool `protobuf:"varint,2,opt,name=missed,proto3" json:"missed,omitempty"`
-}
-
-func (m *MissedBlock) Reset()         { *m = MissedBlock{} }
-func (m *MissedBlock) String() string { return proto.CompactTextString(m) }
-func (*MissedBlock) ProtoMessage()    {}
-func (*MissedBlock) Descriptor() ([]byte, []int) {
-	return fileDescriptor_14205810582f3203, []int{3}
-}
-func (m *MissedBlock) XXX_Unmarshal(b []byte) error {
-	return m.Unmarshal(b)
-}
-func (m *MissedBlock) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	if deterministic {
-		return xxx_messageInfo_MissedBlock.Marshal(b, m, deterministic)
-	} else {
-		b = b[:cap(b)]
-		n, err := m.MarshalToSizedBuffer(b)
-		if err != nil {
-			return nil, err
-		}
-		return b[:n], nil
-	}
-}
-func (m *MissedBlock) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_MissedBlock.Merge(m, src)
-}
-func (m *MissedBlock) XXX_Size() int {
-	return m.Size()
-}
-func (m *MissedBlock) XXX_DiscardUnknown() {
-	xxx_messageInfo_MissedBlock.DiscardUnknown(m)
-}
-
-var xxx_messageInfo_MissedBlock proto.InternalMessageInfo
-
-func (m *MissedBlock) GetIndex() int64 {
-	if m != nil {
-		return m.Index
-	}
-	return 0
-}
-
-func (m *MissedBlock) GetMissed() bool {
-	if m != nil {
-		return m.Missed
-	}
-	return false
-}
-
 func init() {
 	proto.RegisterType((*GenesisState)(nil), "kira.slashing.GenesisState")
 	proto.RegisterType((*SigningInfo)(nil), "kira.slashing.SigningInfo")
-	proto.RegisterType((*ValidatorMissedBlocks)(nil), "kira.slashing.ValidatorMissedBlocks")
-	proto.RegisterType((*MissedBlock)(nil), "kira.slashing.MissedBlock")
 }
 
 func init() { proto.RegisterFile("genesis.proto", fileDescriptor_14205810582f3203) }
 
 var fileDescriptor_14205810582f3203 = []byte{
-	// 407 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x94, 0x92, 0xc1, 0xca, 0xd3, 0x40,
-	0x14, 0x85, 0x33, 0xad, 0x56, 0x9d, 0x34, 0x2e, 0x86, 0xb4, 0x84, 0xa2, 0x69, 0x89, 0x0a, 0xc5,
-	0x45, 0x02, 0xed, 0x4e, 0x77, 0x51, 0x10, 0x11, 0x41, 0x52, 0x70, 0x21, 0x94, 0x32, 0x6d, 0xa6,
-	0xd3, 0xa1, 0x49, 0xa6, 0x64, 0x62, 0x69, 0x77, 0xee, 0xdd, 0xf4, 0x39, 0x7c, 0x92, 0x2e, 0xbb,
-	0x74, 0x55, 0xa4, 0x7d, 0x03, 0x9f, 0x40, 0x3a, 0xd3, 0x60, 0x1a, 0xa2, 0xf0, 0xef, 0x72, 0xb9,
-	0xe7, 0x9c, 0xfb, 0xe5, 0x24, 0xd0, 0xa0, 0x24, 0x21, 0x82, 0x09, 0x77, 0x95, 0xf2, 0x8c, 0x23,
-	0x63, 0xc9, 0x52, 0xec, 0x8a, 0x08, 0x8b, 0x05, 0x4b, 0x68, 0xc7, 0xa4, 0x9c, 0x72, 0xb9, 0xf1,
-	0x2e, 0x4f, 0x4a, 0xd4, 0x79, 0x9c, 0xef, 0xd5, 0xec, 0x7c, 0xaf, 0xc1, 0xe6, 0x3b, 0x15, 0x33,
-	0xca, 0x70, 0x46, 0xd0, 0x10, 0x36, 0x56, 0x38, 0xc5, 0xb1, 0xb0, 0x40, 0x0f, 0xf4, 0xf5, 0x41,
-	0xcb, 0xbd, 0x89, 0x75, 0x3f, 0xc9, 0xa5, 0x7f, 0x6f, 0x7f, 0xec, 0x6a, 0xc1, 0x55, 0x8a, 0xc6,
-	0xd0, 0x10, 0x8c, 0x26, 0x2c, 0xa1, 0x13, 0x96, 0xcc, 0xb9, 0xb0, 0x6a, 0xbd, 0x7a, 0x5f, 0x1f,
-	0x74, 0x4a, 0xde, 0x91, 0xd2, 0xbc, 0x4f, 0xe6, 0xdc, 0x7f, 0x72, 0x09, 0xf8, 0x7d, 0xec, 0x9a,
-	0x5b, 0x1c, 0x47, 0xaf, 0x9c, 0x1b, 0xbb, 0x13, 0x34, 0xc5, 0x5f, 0xa9, 0x40, 0x14, 0x1a, 0x31,
-	0x13, 0x82, 0x84, 0x93, 0x69, 0xc4, 0x67, 0x4b, 0x61, 0xd5, 0x65, 0xfc, 0xf3, 0x52, 0xfc, 0x67,
-	0x1c, 0xb1, 0x10, 0x67, 0x3c, 0xfd, 0x28, 0xc5, 0xbe, 0xd4, 0x96, 0x0f, 0xdd, 0x04, 0x39, 0x41,
-	0x33, 0x2e, 0x68, 0x9d, 0x1f, 0x00, 0xea, 0x05, 0x48, 0x64, 0xc1, 0x07, 0x38, 0x0c, 0x53, 0x22,
-	0x54, 0x1b, 0x8f, 0x82, 0x7c, 0x44, 0xdf, 0x00, 0x6c, 0xaf, 0xf3, 0x7b, 0x93, 0x22, 0xbd, 0x55,
-	0x93, 0xbd, 0x3d, 0xfb, 0x17, 0x5c, 0xb1, 0x84, 0x17, 0x57, 0xb6, 0xa7, 0x8a, 0xad, 0x3a, 0xd0,
-	0x09, 0xcc, 0x75, 0x85, 0xd9, 0xd9, 0x01, 0xd8, 0xaa, 0x7c, 0xe5, 0xff, 0x60, 0x8f, 0xcb, 0x4d,
-	0x56, 0x7f, 0xa8, 0x42, 0xda, 0x9d, 0xfa, 0x7b, 0x0d, 0xf5, 0x82, 0x15, 0x99, 0xf0, 0x3e, 0x4b,
-	0x42, 0xb2, 0x91, 0x14, 0xf5, 0x40, 0x0d, 0xa8, 0x0d, 0x1b, 0xca, 0x24, 0x9b, 0x7a, 0x18, 0x5c,
-	0x27, 0xff, 0xed, 0xfe, 0x64, 0x83, 0xc3, 0xc9, 0x06, 0xbf, 0x4e, 0x36, 0xd8, 0x9d, 0x6d, 0xed,
-	0x70, 0xb6, 0xb5, 0x9f, 0x67, 0x5b, 0xfb, 0xf2, 0x92, 0xb2, 0x6c, 0xf1, 0x75, 0xea, 0xce, 0x78,
-	0xec, 0x7d, 0x60, 0x29, 0x7e, 0xc3, 0x53, 0xe2, 0x09, 0xb2, 0xc4, 0xcc, 0xdb, 0x78, 0x39, 0xb4,
-	0x97, 0x6d, 0x57, 0x44, 0x4c, 0x1b, 0xf2, 0xbf, 0x1e, 0xfe, 0x09, 0x00, 0x00, 0xff, 0xff, 0x40,
-	0xbf, 0x82, 0xd3, 0x1d, 0x03, 0x00, 0x00,
+	// 316 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xe2, 0xe2, 0x4d, 0x4f, 0xcd, 0x4b,
+	0x2d, 0xce, 0x2c, 0xd6, 0x2b, 0x28, 0xca, 0x2f, 0xc9, 0x17, 0xe2, 0xcd, 0xce, 0x2c, 0x4a, 0xd4,
+	0x2b, 0xce, 0x49, 0x2c, 0xce, 0xc8, 0xcc, 0x4b, 0x97, 0x12, 0x49, 0xcf, 0x4f, 0xcf, 0x07, 0xcb,
+	0xe8, 0x83, 0x58, 0x10, 0x45, 0x52, 0x7c, 0x30, 0x79, 0x08, 0x5f, 0x69, 0x11, 0x23, 0x17, 0x8f,
+	0x3b, 0xc4, 0x98, 0xe0, 0x92, 0xc4, 0x92, 0x54, 0x21, 0x63, 0x2e, 0xb6, 0x82, 0xc4, 0xa2, 0xc4,
+	0xdc, 0x62, 0x09, 0x46, 0x05, 0x46, 0x0d, 0x6e, 0x23, 0x51, 0x3d, 0x14, 0x63, 0xf5, 0x02, 0xc0,
+	0x92, 0x4e, 0x2c, 0x27, 0xee, 0xc9, 0x33, 0x04, 0x41, 0x95, 0x0a, 0xc5, 0x72, 0xf1, 0x16, 0x67,
+	0xa6, 0xe7, 0x65, 0xe6, 0xa5, 0xc7, 0x67, 0xe6, 0xa5, 0xe5, 0x17, 0x4b, 0x30, 0x29, 0x30, 0x6b,
+	0x70, 0x1b, 0x49, 0xa1, 0xe9, 0x0d, 0x86, 0xa8, 0xf1, 0xcc, 0x4b, 0xcb, 0x77, 0x92, 0x01, 0x19,
+	0xf0, 0xe9, 0x9e, 0xbc, 0x48, 0x65, 0x62, 0x6e, 0x8e, 0x95, 0x12, 0x8a, 0x76, 0xa5, 0x20, 0x9e,
+	0x62, 0x84, 0xd2, 0x62, 0xa5, 0x55, 0x8c, 0x5c, 0xdc, 0x48, 0x7a, 0x85, 0x24, 0xb8, 0xd8, 0x13,
+	0x53, 0x52, 0x8a, 0x52, 0x8b, 0x21, 0x8e, 0xe4, 0x0c, 0x82, 0x71, 0x85, 0x1a, 0x18, 0xb9, 0xc4,
+	0xca, 0x12, 0x73, 0x32, 0x53, 0x12, 0x4b, 0xf2, 0x8b, 0xe2, 0x91, 0x0d, 0x95, 0x60, 0x02, 0x7b,
+	0x47, 0x19, 0xcd, 0x49, 0x61, 0x30, 0xc5, 0xc8, 0x6e, 0x53, 0x85, 0xba, 0x4d, 0x16, 0xe2, 0x36,
+	0xec, 0x06, 0x2a, 0x05, 0x89, 0x94, 0x61, 0xd3, 0xec, 0x72, 0xe2, 0x91, 0x1c, 0xe3, 0x85, 0x47,
+	0x72, 0x8c, 0x0f, 0x1e, 0xc9, 0x31, 0x4e, 0x78, 0x2c, 0xc7, 0x70, 0xe1, 0xb1, 0x1c, 0xc3, 0x8d,
+	0xc7, 0x72, 0x0c, 0x51, 0x5a, 0xe9, 0x99, 0x25, 0x19, 0xa5, 0x49, 0x7a, 0xc9, 0xf9, 0xb9, 0xfa,
+	0xde, 0x99, 0x45, 0x89, 0xce, 0xf9, 0x45, 0xa9, 0xfa, 0xc5, 0xa9, 0xd9, 0x89, 0x99, 0xfa, 0x15,
+	0xfa, 0x30, 0x17, 0xe9, 0x97, 0x54, 0x16, 0xa4, 0x16, 0x27, 0xb1, 0x81, 0xa3, 0xc7, 0x18, 0x10,
+	0x00, 0x00, 0xff, 0xff, 0xc1, 0x72, 0x19, 0xa3, 0xe4, 0x01, 0x00, 0x00,
 }
 
 func (m *GenesisState) Marshal() (dAtA []byte, err error) {
@@ -314,20 +185,6 @@ func (m *GenesisState) MarshalToSizedBuffer(dAtA []byte) (int, error) {
 	_ = i
 	var l int
 	_ = l
-	if len(m.MissedBlocks) > 0 {
-		for iNdEx := len(m.MissedBlocks) - 1; iNdEx >= 0; iNdEx-- {
-			{
-				size, err := m.MissedBlocks[iNdEx].MarshalToSizedBuffer(dAtA[:i])
-				if err != nil {
-					return 0, err
-				}
-				i -= size
-				i = encodeVarintGenesis(dAtA, i, uint64(size))
-			}
-			i--
-			dAtA[i] = 0x1a
-		}
-	}
 	if len(m.SigningInfos) > 0 {
 		for iNdEx := len(m.SigningInfos) - 1; iNdEx >= 0; iNdEx-- {
 			{
@@ -395,88 +252,6 @@ func (m *SigningInfo) MarshalToSizedBuffer(dAtA []byte) (int, error) {
 	return len(dAtA) - i, nil
 }
 
-func (m *ValidatorMissedBlocks) Marshal() (dAtA []byte, err error) {
-	size := m.Size()
-	dAtA = make([]byte, size)
-	n, err := m.MarshalToSizedBuffer(dAtA[:size])
-	if err != nil {
-		return nil, err
-	}
-	return dAtA[:n], nil
-}
-
-func (m *ValidatorMissedBlocks) MarshalTo(dAtA []byte) (int, error) {
-	size := m.Size()
-	return m.MarshalToSizedBuffer(dAtA[:size])
-}
-
-func (m *ValidatorMissedBlocks) MarshalToSizedBuffer(dAtA []byte) (int, error) {
-	i := len(dAtA)
-	_ = i
-	var l int
-	_ = l
-	if len(m.MissedBlocks) > 0 {
-		for iNdEx := len(m.MissedBlocks) - 1; iNdEx >= 0; iNdEx-- {
-			{
-				size, err := m.MissedBlocks[iNdEx].MarshalToSizedBuffer(dAtA[:i])
-				if err != nil {
-					return 0, err
-				}
-				i -= size
-				i = encodeVarintGenesis(dAtA, i, uint64(size))
-			}
-			i--
-			dAtA[i] = 0x12
-		}
-	}
-	if len(m.Address) > 0 {
-		i -= len(m.Address)
-		copy(dAtA[i:], m.Address)
-		i = encodeVarintGenesis(dAtA, i, uint64(len(m.Address)))
-		i--
-		dAtA[i] = 0xa
-	}
-	return len(dAtA) - i, nil
-}
-
-func (m *MissedBlock) Marshal() (dAtA []byte, err error) {
-	size := m.Size()
-	dAtA = make([]byte, size)
-	n, err := m.MarshalToSizedBuffer(dAtA[:size])
-	if err != nil {
-		return nil, err
-	}
-	return dAtA[:n], nil
-}
-
-func (m *MissedBlock) MarshalTo(dAtA []byte) (int, error) {
-	size := m.Size()
-	return m.MarshalToSizedBuffer(dAtA[:size])
-}
-
-func (m *MissedBlock) MarshalToSizedBuffer(dAtA []byte) (int, error) {
-	i := len(dAtA)
-	_ = i
-	var l int
-	_ = l
-	if m.Missed {
-		i--
-		if m.Missed {
-			dAtA[i] = 1
-		} else {
-			dAtA[i] = 0
-		}
-		i--
-		dAtA[i] = 0x10
-	}
-	if m.Index != 0 {
-		i = encodeVarintGenesis(dAtA, i, uint64(m.Index))
-		i--
-		dAtA[i] = 0x8
-	}
-	return len(dAtA) - i, nil
-}
-
 func encodeVarintGenesis(dAtA []byte, offset int, v uint64) int {
 	offset -= sovGenesis(v)
 	base := offset
@@ -502,12 +277,6 @@ func (m *GenesisState) Size() (n int) {
 			n += 1 + l + sovGenesis(uint64(l))
 		}
 	}
-	if len(m.MissedBlocks) > 0 {
-		for _, e := range m.MissedBlocks {
-			l = e.Size()
-			n += 1 + l + sovGenesis(uint64(l))
-		}
-	}
 	return n
 }
 
@@ -523,40 +292,6 @@ func (m *SigningInfo) Size() (n int) {
 	}
 	l = m.ValidatorSigningInfo.Size()
 	n += 1 + l + sovGenesis(uint64(l))
-	return n
-}
-
-func (m *ValidatorMissedBlocks) Size() (n int) {
-	if m == nil {
-		return 0
-	}
-	var l int
-	_ = l
-	l = len(m.Address)
-	if l > 0 {
-		n += 1 + l + sovGenesis(uint64(l))
-	}
-	if len(m.MissedBlocks) > 0 {
-		for _, e := range m.MissedBlocks {
-			l = e.Size()
-			n += 1 + l + sovGenesis(uint64(l))
-		}
-	}
-	return n
-}
-
-func (m *MissedBlock) Size() (n int) {
-	if m == nil {
-		return 0
-	}
-	var l int
-	_ = l
-	if m.Index != 0 {
-		n += 1 + sovGenesis(uint64(m.Index))
-	}
-	if m.Missed {
-		n += 2
-	}
 	return n
 }
 
@@ -659,40 +394,6 @@ func (m *GenesisState) Unmarshal(dAtA []byte) error {
 			}
 			m.SigningInfos = append(m.SigningInfos, SigningInfo{})
 			if err := m.SigningInfos[len(m.SigningInfos)-1].Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
-				return err
-			}
-			iNdEx = postIndex
-		case 3:
-			if wireType != 2 {
-				return fmt.Errorf("proto: wrong wireType = %d for field MissedBlocks", wireType)
-			}
-			var msglen int
-			for shift := uint(0); ; shift += 7 {
-				if shift >= 64 {
-					return ErrIntOverflowGenesis
-				}
-				if iNdEx >= l {
-					return io.ErrUnexpectedEOF
-				}
-				b := dAtA[iNdEx]
-				iNdEx++
-				msglen |= int(b&0x7F) << shift
-				if b < 0x80 {
-					break
-				}
-			}
-			if msglen < 0 {
-				return ErrInvalidLengthGenesis
-			}
-			postIndex := iNdEx + msglen
-			if postIndex < 0 {
-				return ErrInvalidLengthGenesis
-			}
-			if postIndex > l {
-				return io.ErrUnexpectedEOF
-			}
-			m.MissedBlocks = append(m.MissedBlocks, ValidatorMissedBlocks{})
-			if err := m.MissedBlocks[len(m.MissedBlocks)-1].Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
 				return err
 			}
 			iNdEx = postIndex
@@ -814,217 +515,6 @@ func (m *SigningInfo) Unmarshal(dAtA []byte) error {
 				return err
 			}
 			iNdEx = postIndex
-		default:
-			iNdEx = preIndex
-			skippy, err := skipGenesis(dAtA[iNdEx:])
-			if err != nil {
-				return err
-			}
-			if skippy < 0 {
-				return ErrInvalidLengthGenesis
-			}
-			if (iNdEx + skippy) < 0 {
-				return ErrInvalidLengthGenesis
-			}
-			if (iNdEx + skippy) > l {
-				return io.ErrUnexpectedEOF
-			}
-			iNdEx += skippy
-		}
-	}
-
-	if iNdEx > l {
-		return io.ErrUnexpectedEOF
-	}
-	return nil
-}
-func (m *ValidatorMissedBlocks) Unmarshal(dAtA []byte) error {
-	l := len(dAtA)
-	iNdEx := 0
-	for iNdEx < l {
-		preIndex := iNdEx
-		var wire uint64
-		for shift := uint(0); ; shift += 7 {
-			if shift >= 64 {
-				return ErrIntOverflowGenesis
-			}
-			if iNdEx >= l {
-				return io.ErrUnexpectedEOF
-			}
-			b := dAtA[iNdEx]
-			iNdEx++
-			wire |= uint64(b&0x7F) << shift
-			if b < 0x80 {
-				break
-			}
-		}
-		fieldNum := int32(wire >> 3)
-		wireType := int(wire & 0x7)
-		if wireType == 4 {
-			return fmt.Errorf("proto: ValidatorMissedBlocks: wiretype end group for non-group")
-		}
-		if fieldNum <= 0 {
-			return fmt.Errorf("proto: ValidatorMissedBlocks: illegal tag %d (wire type %d)", fieldNum, wire)
-		}
-		switch fieldNum {
-		case 1:
-			if wireType != 2 {
-				return fmt.Errorf("proto: wrong wireType = %d for field Address", wireType)
-			}
-			var stringLen uint64
-			for shift := uint(0); ; shift += 7 {
-				if shift >= 64 {
-					return ErrIntOverflowGenesis
-				}
-				if iNdEx >= l {
-					return io.ErrUnexpectedEOF
-				}
-				b := dAtA[iNdEx]
-				iNdEx++
-				stringLen |= uint64(b&0x7F) << shift
-				if b < 0x80 {
-					break
-				}
-			}
-			intStringLen := int(stringLen)
-			if intStringLen < 0 {
-				return ErrInvalidLengthGenesis
-			}
-			postIndex := iNdEx + intStringLen
-			if postIndex < 0 {
-				return ErrInvalidLengthGenesis
-			}
-			if postIndex > l {
-				return io.ErrUnexpectedEOF
-			}
-			m.Address = string(dAtA[iNdEx:postIndex])
-			iNdEx = postIndex
-		case 2:
-			if wireType != 2 {
-				return fmt.Errorf("proto: wrong wireType = %d for field MissedBlocks", wireType)
-			}
-			var msglen int
-			for shift := uint(0); ; shift += 7 {
-				if shift >= 64 {
-					return ErrIntOverflowGenesis
-				}
-				if iNdEx >= l {
-					return io.ErrUnexpectedEOF
-				}
-				b := dAtA[iNdEx]
-				iNdEx++
-				msglen |= int(b&0x7F) << shift
-				if b < 0x80 {
-					break
-				}
-			}
-			if msglen < 0 {
-				return ErrInvalidLengthGenesis
-			}
-			postIndex := iNdEx + msglen
-			if postIndex < 0 {
-				return ErrInvalidLengthGenesis
-			}
-			if postIndex > l {
-				return io.ErrUnexpectedEOF
-			}
-			m.MissedBlocks = append(m.MissedBlocks, MissedBlock{})
-			if err := m.MissedBlocks[len(m.MissedBlocks)-1].Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
-				return err
-			}
-			iNdEx = postIndex
-		default:
-			iNdEx = preIndex
-			skippy, err := skipGenesis(dAtA[iNdEx:])
-			if err != nil {
-				return err
-			}
-			if skippy < 0 {
-				return ErrInvalidLengthGenesis
-			}
-			if (iNdEx + skippy) < 0 {
-				return ErrInvalidLengthGenesis
-			}
-			if (iNdEx + skippy) > l {
-				return io.ErrUnexpectedEOF
-			}
-			iNdEx += skippy
-		}
-	}
-
-	if iNdEx > l {
-		return io.ErrUnexpectedEOF
-	}
-	return nil
-}
-func (m *MissedBlock) Unmarshal(dAtA []byte) error {
-	l := len(dAtA)
-	iNdEx := 0
-	for iNdEx < l {
-		preIndex := iNdEx
-		var wire uint64
-		for shift := uint(0); ; shift += 7 {
-			if shift >= 64 {
-				return ErrIntOverflowGenesis
-			}
-			if iNdEx >= l {
-				return io.ErrUnexpectedEOF
-			}
-			b := dAtA[iNdEx]
-			iNdEx++
-			wire |= uint64(b&0x7F) << shift
-			if b < 0x80 {
-				break
-			}
-		}
-		fieldNum := int32(wire >> 3)
-		wireType := int(wire & 0x7)
-		if wireType == 4 {
-			return fmt.Errorf("proto: MissedBlock: wiretype end group for non-group")
-		}
-		if fieldNum <= 0 {
-			return fmt.Errorf("proto: MissedBlock: illegal tag %d (wire type %d)", fieldNum, wire)
-		}
-		switch fieldNum {
-		case 1:
-			if wireType != 0 {
-				return fmt.Errorf("proto: wrong wireType = %d for field Index", wireType)
-			}
-			m.Index = 0
-			for shift := uint(0); ; shift += 7 {
-				if shift >= 64 {
-					return ErrIntOverflowGenesis
-				}
-				if iNdEx >= l {
-					return io.ErrUnexpectedEOF
-				}
-				b := dAtA[iNdEx]
-				iNdEx++
-				m.Index |= int64(b&0x7F) << shift
-				if b < 0x80 {
-					break
-				}
-			}
-		case 2:
-			if wireType != 0 {
-				return fmt.Errorf("proto: wrong wireType = %d for field Missed", wireType)
-			}
-			var v int
-			for shift := uint(0); ; shift += 7 {
-				if shift >= 64 {
-					return ErrIntOverflowGenesis
-				}
-				if iNdEx >= l {
-					return io.ErrUnexpectedEOF
-				}
-				b := dAtA[iNdEx]
-				iNdEx++
-				v |= int(b&0x7F) << shift
-				if b < 0x80 {
-					break
-				}
-			}
-			m.Missed = bool(v != 0)
 		default:
 			iNdEx = preIndex
 			skippy, err := skipGenesis(dAtA[iNdEx:])