+++ x/staking/keeper/slash.go
@@ -4,6 +4,7 @@ import (
 	"github.com/KiraCore/sekai/x/staking/types"
 	customstakingtypes "github.com/KiraCore/sekai/x/staking/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 )
 
 // Activate a validator
@@ -14,7 +15,7 @@ func (k Keeper) Activate(ctx sdk.Context, valAddress sdk.ValAddress) error {
 	}
 
 	if validator.IsPaused() {
-		return customstakingtypes.ErrValidatorPaused
+		return sdkerrors.Wrap(customstakingtypes.ErrValidatorPaused, "Can NOT activate paused validator, you must unpause")
 	}
 
 	k.setStatusToValidator(ctx, validator, customstakingtypes.Active)
@@ -22,6 +23,18 @@ func (k Keeper) Activate(ctx sdk.Context, valAddress sdk.ValAddress) error {
 	k.RemoveRemovingValidator(ctx, validator)
 
 	return nil
+}
+
+// ResetWholeValidatorRank reset whole validators' status, rank and streak
+func (k Keeper) ResetWholeValidatorRank(ctx sdk.Context) {
+	// TODO: is it correct to use this iterator @Jonathan?
+	k.IterateValidators(ctx, func(index int64, validator *types.Validator) (stop bool) {
+		validator.Status = customstakingtypes.Active
+		validator.Rank = 0
+		validator.Streak = 0
+		k.AddValidator(ctx, *validator)
+		return false
+	})
 }
 
 // Inactivate inactivate the validator