+++ x/tokens/handler_test.go
@@ -298,6 +298,8 @@ func TestHandler_CreateProposalUpsertTokenAliases(t *testing.T) {
 		ctx.BlockTime().Add(time.Second*time.Duration(properties.ProposalEndTime)+
 			time.Second*time.Duration(properties.ProposalEnactmentTime),
 		),
+		ctx.BlockHeight()+2,
+		ctx.BlockHeight()+3,
 		"some desc",
 	)
 	require.NoError(t, err)
@@ -404,6 +406,8 @@ func TestHandler_CreateProposalUpsertTokenRates(t *testing.T) {
 		ctx.BlockTime().Add(time.Second*time.Duration(properties.ProposalEndTime)+
 			time.Second*time.Duration(properties.ProposalEnactmentTime),
 		),
+		ctx.BlockHeight()+2,
+		ctx.BlockHeight()+3,
 		"some desc",
 	)
 	require.NoError(t, err)